+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Loading Spinners
 - Footer
+- more search fields in Event search filter
 
 ### Changed